Fixed benchmark calculations, was overflowing on 32bit.

+++ benchmark/benchmark.c
@@ -33,13 +33,17 @@ void benchmark_check_time(gearman_benchmark_st *benchmark)
   { 
     benchmark->total_jobs+= benchmark->jobs;
 
-    printf("[Current: %6u jobs/s, Total: %6u jobs/s]\n",
-           (benchmark->jobs * 1000000) /
-           (((benchmark->end.tv_sec * 1000000) + benchmark->end.tv_usec) -
-            ((benchmark->begin.tv_sec * 1000000) + benchmark->begin.tv_usec)),
-           (benchmark->total_jobs * 1000000) /
-           (((benchmark->end.tv_sec * 1000000) + benchmark->end.tv_usec) -
-            ((benchmark->total.tv_sec * 1000000) + benchmark->total.tv_usec)));
+    printf("[Current: %6"PRIu64" jobs/s, Total: %6"PRIu64" jobs/s]\n",
+           ((uint64_t)(benchmark->jobs) * 1000000) /
+           ((((uint64_t)(benchmark->end.tv_sec) * 1000000) +
+              (uint64_t)(benchmark->end.tv_usec)) -
+            (((uint64_t)(benchmark->begin.tv_sec) * 1000000) +
+              (uint64_t)(benchmark->begin.tv_usec))),
+           ((uint64_t)(benchmark->total_jobs) * 1000000) /
+           ((((uint64_t)(benchmark->end.tv_sec) * 1000000) +
+              (uint64_t)(benchmark->end.tv_usec)) -
+            (((uint64_t)(benchmark->total.tv_sec) * 1000000) +
+              (uint64_t)(benchmark->total.tv_usec))));
 
     benchmark->jobs= 0;
     gettimeofday(&(benchmark->begin), NULL);
